// github/types/paths/notifications/threads/{thread_id}/subscription.types.ts
export type activityGetThreadSubscriptionForAuthenticatedUser = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: activityGetThreadSubscriptionForAuthenticatedUser_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: thread_subscription;
          };
        };
        examples: {
          default: unknown;
        };
      };
      304: not_modified;
      401: requires_authentication;
      403: forbidden;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;
// ...
export type activitySetThreadSubscription = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: activitySetThreadSubscription_Path;
    headers: never;
    cookie: never;
    body: {
      /**
       * Whether to block all notifications from a thread.
       * @default false
       */
      ignored?: boolean;
    };
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: thread_subscription;
          };
        };
        examples: {
          default: unknown;
        };
      };
      304: not_modified;
      401: requires_authentication;
      403: forbidden;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;
// ...
export type activityDeleteThreadSubscription = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: activityDeleteThreadSubscription_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      304: not_modified;
      401: requires_authentication;
      403: forbidden;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type activityGetThreadSubscriptionForAuthenticatedUser_Path = {
  thread_id: number;
};

export type activitySetThreadSubscription_Path = { thread_id: number };

export type activityDeleteThreadSubscription_Path = { thread_id: number };